add location listener and Use Hilt to Inject Services to Classes

+++ app/src/main/java/com/example/dvtweather/Webservice.kt
@@ -8,7 +8,7 @@ const val appID: String = "" // Add API KEY
 interface Webservice {
 
     @GET("/data/2.5/weather?lat={latitude}&lon={longitude}&appid=$appID")
-    suspend fun getWeather(@Path("weather") latitude: Long, longitude:Long): Weather
+    suspend fun getWeather(@Path("weather") latitude: Double, longitude:Double): Weather
 
     @GET("/data/2.5/forecast?q={city}&appid=$appID")
     suspend fun getForecast(@Path("forecast") city: String): Forecast
